Add files via upload
This project is a Monks and Monsters River Crossing game built in C++. The goal is to safely transfer all monks and monsters from the right side of the river to the left using a boat, while ensuring that monsters never outnumber monks on either side. It involves logical thinking and user interaction to win the game without getting the monks eaten.